LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-en-GB-expected.txt

PASS navigator.language is "en-gb"
PASS successfullyParsed is true

TEST COMPLETE
